moving units out of procedures and device
