+++ src/aind_data_schema/base.py
@@ -66,6 +66,7 @@ class Registry:
     """
     Class to store common registries for use in PIDNames
     """
+
     ROR = BaseName(name="Research Organization Registry", abbreviation="ROR")
     NCBI = BaseName(name="National Center for Biotechnology Information", abbreviation="NCBI")
